+++ lib/src/dart/typedef.dart
@@ -1,3 +1,2 @@
 typedef FV = Future<void>;
-typedef FB = Future<bool>;
 typedef JSON = Map<String, dynamic>;