+++ lib/src/flutter/extensions.dart
@@ -95,27 +95,6 @@ extension HaloListSpan on List<InlineSpan> {
   }
 }
 
-extension HaloString on String {}
-
-extension HaloTextStyle<S extends TextStyle> on S {
-  S cwc(Color? color) => copyWith(color: color) as S;
-
-  S cww(FontWeight? fontWeight) => copyWith(fontWeight: fontWeight) as S;
-
-  S cws(double? fontSize) => copyWith(fontSize: fontSize) as S;
-
-  S cwff(String? fontFamily) => copyWith(fontFamily: fontFamily) as S;
-
-  S cwh(double? height) => copyWith(height: height) as S;
-}
-
-extension HaloEdgeInsets on EdgeInsets {
-  double get l => left;
-  double get r => right;
-  double get t => top;
-  double get b => bottom;
-}
-
 extension WidgetDebugger on Widget {
   Widget get debug {
     if (!kDebugMode) return this;